Fix existing tests

+++ core/src/main/java/de/muenchen/rbs/kitafinderdatenservice/domain/events/Outboxevent.java
@@ -60,7 +60,7 @@ public class Outboxevent implements Serializable {
 	@Lob
 	private String errorMessage;
 
-	private Long retryCount;
+	private Integer retryCount;
 
 	private LocalDateTime retryNotBefore;
